refactor(auth): dedicated email password module
