+++ packages/@robojs/auth/src/providers/email-password.ts
@@ -1,0 +1,2 @@
+export { default } from '../builtins/email-password/index.js'
+export * from '../builtins/email-password/index.js'
